Add base server implementation
Added a base server implemenatation. This class can be extended
instead of implement the server interface directly.

The base implementation is taking care of all the
parts that always occurs when implementing a new protocol.

Examples of this is cleanup, adding contexts, remove
suppliers and all the synhronization around these parts.

This means that it's very little you need to implement
on the server. However, you still need to define what makes
a context unique, if any additional server cleanup should
be done, and how the protocol context should be created.